Built Profile Page
I put on a profile page, but I am not happy about the design of it. I
like how it is set up, but it will take some Sass wizardry to change it
to something I like.

Also added a js file. Nothing is in it, but there are some additions I
want to do, just not right now.